+++ CHANGELOG.md
@@ -17,3 +17,5 @@
 * Added Q-score comparison for T-free and T-bearing regions
 * T-free visualization will not be shown if no bases fit the T-free criteria
 * Added backend support for analysing bearing and free regions of any base
+* Free regions are mow defined as the neighboring 3 bases to the target instead of 6
+* Added backend histogram visualization for free and bearing regions